Fix implicit conversion warning (#16598)
* Fix implicit conversion warning

* Change SAXParser::textHandler parameter type from int to size_t

+++ cocos/platform/CCSAXParser.cpp
@@ -83,7 +83,7 @@ bool XmlSaxHander::VisitExit( const tinyxml2::XMLElement& element )
 bool XmlSaxHander::Visit( const tinyxml2::XMLText& text )
 {
     //log("Visit %s",text.Value());
-    SAXParser::textHandler(_ccsaxParserImp, (const CC_XML_CHAR *)text.Value(), static_cast<int>(strlen(text.Value())));
+    SAXParser::textHandler(_ccsaxParserImp, (const CC_XML_CHAR *)text.Value(), strlen(text.Value()));
     return true;
 }
 
@@ -163,7 +163,7 @@ bool SAXParser::parseIntrusive(char* xmlData, size_t dataLength)
 
     rapidxml::xml_sax3_parser<> parser(&printer);
     try {
-        parser.parse<>(xmlData, dataLength);
+        parser.parse<>(xmlData, static_cast<int>(dataLength));
         return true;
     }
     catch (rapidxml::parse_error& e)
@@ -184,7 +184,7 @@ void SAXParser::endElement(void *ctx, const CC_XML_CHAR *name)
 {
     ((SAXParser*)(ctx))->_delegator->endElement(ctx, (char*)name);
 }
-void SAXParser::textHandler(void *ctx, const CC_XML_CHAR *name, int len)
+void SAXParser::textHandler(void *ctx, const CC_XML_CHAR *name, size_t len)
 {
     ((SAXParser*)(ctx))->_delegator->textHandler(ctx, (char*)name, len);
 }
